Bảng cv+ bảng skill

+++ asm1/src/app/@core/services/database.js
@@ -21,6 +21,67 @@ const getAll = (table, callback) => {
     callback(err, results);
   });
 };
+
+const getAllCV = (callback) => {
+  const sql = `SELECT 
+  userinfo.*, 
+  language.*, 
+  skill.*, 
+  experience.*, 
+  education.*, 
+  activity.*, 
+  informationtechnologyexperience.*, 
+  certificate.*
+FROM userinfo
+LEFT JOIN language ON userinfo.id = language.user_id
+LEFT JOIN skill ON userinfo.id = skill.user_id
+LEFT JOIN recruitment ON userinfo.id = recruitment.user_id
+LEFT JOIN experience ON userinfo.id = experience.user_id
+LEFT JOIN education ON userinfo.id = education.user_id
+LEFT JOIN activity ON userinfo.id = activity.user_id
+LEFT JOIN informationtechnologyexperience ON userinfo.id = informationtechnologyexperience.user_id
+LEFT JOIN certificate ON userinfo.id = certificate.user_id`;
+  db.query(sql, (err, results) => {
+    callback(err, results);
+  });
+};
+
+
+const getAllCVByID = (callback, id) => {
+  const sql = `
+  SELECT 
+  userinfo.*, 
+  language.*,
+  language.level as language_level,
+  skill.*,
+  recruitment.*,
+  experience.*, 
+  DATE_FORMAT(experience.startdate, '%d/%m/%Y') AS start_date,
+  DATE_FORMAT(experience.enddate, '%d/%m/%Y') AS end_date,
+  education.*,
+  education.name as edu_name,
+  YEAR(education.startTime) AS start_year,
+  YEAR(education.endTime) AS end_year,
+  activity.*, 
+  informationtechnologyexperience.*, 
+  informationtechnologyexperience.level as info_level,
+  certificate.*
+FROM userinfo
+LEFT JOIN language ON userinfo.id = language.user_id
+LEFT JOIN skill ON userinfo.id = skill.user_id
+LEFT JOIN recruitment ON userinfo.id = recruitment.user_id
+LEFT JOIN experience ON userinfo.id = experience.user_id
+LEFT JOIN education ON userinfo.id = education.user_id
+LEFT JOIN activity ON userinfo.id = activity.user_id
+LEFT JOIN informationtechnologyexperience ON userinfo.id = informationtechnologyexperience.user_id
+LEFT JOIN certificate ON userinfo.id = certificate.user_id
+WHERE userinfo.id = ?
+`;
+  db.query(sql, [id], (err, results) => {
+    callback(err, results);
+  });
+};
+
 const getByID = (table, id, callback) => {
   const sql = `SELECT * FROM ?? WHERE ?`;
   db.query(sql, [table, id], (err, results) => {
@@ -43,25 +104,18 @@ const update = (table, data, id, callback) => {
 };
 
 const Delete = (table, id, callback) => {
-  const sql = `DELETE FROM ?? WHERE ?`;
+  const sql = `DELETE FROM ?? WHERE  ?`;
   db.query(sql, [table, id], (err, results) => {
     callback(err, results);
   });
 };
-const getAllSkillsByUserId = (callback) => {
-  const sql = `
-    SELECT 
-        user_id,
-        GROUP_CONCAT(skill SEPARATOR ',') AS kynang
-    FROM
-        skill
-    GROUP BY user_id
-  `;
 
-  db.query(sql, (err, results) => {
+const DeleteSkill = (callback, id) => {
+  const sql = `DELETE FROM skill WHERE user_id = ?`;
+  db.query(sql, [id], (err, results) => {
     callback(err, results);
   });
-}
+};
 
 
 module.exports = {
@@ -70,6 +124,8 @@ module.exports = {
   update,
   Delete,
   getByID,
-  getAllSkillsByUserId,
+  getAllCV,
+  getAllCVByID,
+  DeleteSkill,
   db
 };
